% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_pathway}
\alias{analyze_pathway}
\title{Final Network Pathway Analysis}
\usage{
analyze_pathway(df_net_sub, lst_net, cell_type_a, cell_type_b, beta, ntrial)
}
\arguments{
\item{df_net_sub}{A subset of the final network (pathway); for example, the
output of the \code{extract_pathways} function}

\item{lst_net}{Integrated network}

\item{cell_type_a}{Name of cell type A that matches scRNA-seq file; for
example, \code{"Fibroblasts"}}

\item{cell_type_b}{Name of cell type B that matches scRNA-seq file; for
example, \code{"LuminalEpithelialCells"}}

\item{beta}{Upper limit of the test values of the PCSF objective function
parameter $I^2$, which is inversely proportional to the total number of
genes in a given cell-type pair; suggested to be 100 (default) if the
total number of genes in a given cell-type pair is above 10,000; if the
total number of genes is below 5,000, increase to 500}

\item{ntrial}{How many empirical simulations to run? (Sample used to form
theoretical Gamma distribution)}
}
\value{
A data-frame containing information relating to pathway size, mean
node prize, mean edge cost, potential scores, and p-values from a fitted
Gamma distribution
}
\description{
Final Network Pathway Analysis
}
\examples{
{
df_net_sub <- result_cyto$pathways$raw[[1]]
lst_net <- result_cyto$integrated_net
cell_type_a <- "Macrophages"
cell_type_b <- "LuminalEpithelialCells"
beta <- 20
ntrial <- 1000
result <- analyze_pathway(df_net_sub, lst_net, cell_type_a, cell_type_b, beta, ntrial)
}

}
