% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/mutinfo.R
\name{doc_mutinfo}
\alias{doc_mutinfo}
\alias{discretize_sparse}
\alias{mutinfo_xy}
\alias{mi_mat_parallel}
\title{Information Theory}
\usage{
discretize_sparse(mat, disc = "equalfreq", nbins = nrow(mat)^(1/3))

mutinfo_xy(x, y, method = "emp", normalize = FALSE)

mi_mat_parallel(mat, method = "emp", normalize = FALSE)
}
\arguments{
\item{mat}{Any discrete numerical matrix, processed column-wise}

\item{disc}{Name of discretization method. Options include "equalfreq",
"equalwidth", and "globalequalwidth". See \code{infotheo::discretize} for more
detail}

\item{nbins}{The number of bins by which the continuous data should be
discretized; by default the cube root of the number of samples}

\item{x, y}{Any discrete numerical vector}

\item{method}{Name of entropy estimator. Options include "emp", "mm",
"shrink", and "sg". See \code{infotheo::mutinformation} for more detail}

\item{normalize}{Should the entropies be normalized to fall within the range
[0, 1]?}
}
\value{
Mutual information calculations for filtered matrices
}
\description{
Information Theory
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
pcg <- CytoTalk::pcg_human
cell_type_a <- "Macrophages"
cutoff_a <- 0.8
mat_a <- extract_group(cell_type_a, lst_scrna)
mat_filt_a <- subset_rownames(subset_non_zero(mat_a, cutoff_a), pcg)
mat_disc_a <- discretize_sparse(Matrix::t(mat_filt_a))
result <- mi_mat_parallel(mat_disc_a, method = "mm")
}

}
