% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/pcst.R
\name{doc_pcst}
\alias{doc_pcst}
\alias{run_pcst}
\alias{summarize_pcst}
\alias{ks_test_pcst}
\alias{extract_network}
\alias{extract_pathways}
\title{Prize-Collecting Steiner Tree and Pathway Extraction}
\usage{
run_pcst(lst_net, beta_max, omega_min, omega_max)

summarize_pcst(lst_pcst)

ks_test_pcst(lst_pcst)

extract_network(lst_net, lst_pcst, mat_pem, beta, omega)

extract_pathways(df_net, cell_type_a, depth)
}
\arguments{
\item{lst_net}{Integrated network}

\item{beta_max}{Upper limit of the test values of the PCSF objective
function parameter $I^2$, which is inversely proportional to the total
number of genes in a given cell-type pair; suggested to be 100 (default)
if the total number of genes in a given cell-type pair is above 10,000; if
the total number of genes is below 5,000, increase to 500}

\item{omega_min}{Start point of omega range; omega represents the edge cost
of the artificial network, but has been found to be less significant than
beta. Recommended minimum of \code{0.5}}

\item{omega_max}{End point of range between \code{omega_min} and \code{omega_max},
step size of \code{0.1}. Recommended maximum of \code{1.5}}

\item{lst_pcst}{PCST output}

\item{mat_pem}{PEM output}

\item{beta}{A single beta value, see \code{beta_max} paramter for more detail}

\item{omega}{A single omega value, see \code{omega_min} for more detail}

\item{df_net}{Final network; for example, the output of the
\code{extract_network} function}

\item{cell_type_a}{Name of cell type A that matches scRNA-seq file; for
example, \code{"Fibroblasts"}}

\item{depth}{Starting at each ligand-receptor pair in the resultant network,
how many steps out from that pair should be taken to generate each
neighborhood?}
}
\value{
Scores for the integrated network
}
\description{
Prize-Collecting Steiner Tree and Pathway Extraction
}
\examples{
{
lst_net <- CytoTalk::result_cyto$integrated_net
beta_max <- 100
omega_min <- 0.5
omega_max <- 0.5
# result <- run_pcst(lst_net, beta_max, omega_min, omega_max)
}

}
