% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{extract_group}
\alias{extract_group}
\title{Extract Cell Type from Meta List}
\usage{
extract_group(group, lst)
}
\arguments{
\item{group}{Character vector with a single string, name of cell type to
extract}

\item{lst}{A meta matrix list (outputted from \code{doc_fileio} methods)}
}
\value{
A matrix
}
\description{
Extract Cell Type from Meta List
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
cell_type_a <- "Macrophages"
result <- extract_group(cell_type_a, lst_scrna)
}

}
