% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{from_single_cell_experiment}
\alias{from_single_cell_experiment}
\title{Convert SingleCellExperiment to Named List}
\usage{
from_single_cell_experiment(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
A named list containing a sparse data matrix and cell type metadata
}
\description{
Convert SingleCellExperiment to Named List
}
