% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{group_meta}
\alias{group_meta}
\title{Meta List to Named List}
\usage{
group_meta(lst)
}
\arguments{
\item{lst}{A meta matrix list (outputted from \code{doc_fileio} methods)}
}
\value{
A list of matrices
}
\description{
Meta List to Named List
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
result <- group_meta(lst_scrna)
}

}
