% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lrp_mouse}
\alias{lrp_mouse}
\title{Mouse ligand-receptor pairs}
\format{
A data frame with two variables: \code{ligand} and \code{receptor}
}
\usage{
lrp_mouse
}
\description{
Contains the names of mouse ligand-receptor pairs
}
\keyword{datasets}
