% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{match_lr_pairs}
\alias{match_lr_pairs}
\title{Find Ligand Receptor Pairs in Matrix}
\usage{
match_lr_pairs(mat, lrp)
}
\arguments{
\item{mat}{A numerical matrix}

\item{lrp}{Ligand-receptor pair data.frame; see \code{CytoTalk::lrp_human} for
an example}
}
\value{
A matrix with two columns (ligand-receptor)
}
\description{
Find Ligand Receptor Pairs in Matrix
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
lrp <- CytoTalk::lrp_human
cell_type_a <- "Macrophages"
mat_a <- extract_group(cell_type_a, lst_scrna)
result <- match_lr_pairs(mat_a, lrp)
}

}
