% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{read_matrix_folder}
\alias{read_matrix_folder}
\title{Read Folder with scRNAseq Data}
\usage{
read_matrix_folder(
  dpath,
  pattern = ".*scRNAseq_(.+?)\\\\..+",
  auto_transform = TRUE
)
}
\arguments{
\item{dpath}{The path of a directory, which contains scRNAseq matrices}

\item{pattern}{A regular expression, matches scRNAseq filenames}

\item{auto_transform}{Should count data be transformed if detected?}
}
\value{
A named list containing a sparse data matrix and cell type metadata
}
\description{
Read Folder with scRNAseq Data
}
\examples{
{
dir_in <- "~/Tan-Lab/scRNAseq-data"
# lst_scrna <- CytoTalk::read_matrix_folder(dir_in)
# result <- str(lst_scrna$cell_types)
}

}
