% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotalk.R
\name{run_cytotalk}
\alias{run_cytotalk}
\title{Main CytoTalk Pipeline}
\usage{
run_cytotalk(
  lst_scrna,
  cell_type_a,
  cell_type_b,
  cutoff_a = 0.2,
  cutoff_b = 0.2,
  pcg = CytoTalk::pcg_human,
  lrp = CytoTalk::lrp_human,
  beta_max = 100,
  omega_min = 0.5,
  omega_max = 0.5,
  depth = 3,
  ntrial = 1000,
  cores = NULL,
  echo = TRUE,
  dir_out = NULL
)
}
\arguments{
\item{lst_scrna}{List containing scRNA-seq data; for example, lists returned
from \code{read_matrix_folder} or \code{read_matrix_with_meta}}

\item{cell_type_a}{Name of cell type A that matches scRNA-seq file; for
example, \code{"Fibroblasts"}}

\item{cell_type_b}{Name of cell type B that matches scRNA-seq file; for
example, \code{"LuminalEpithelialCells"}}

\item{cutoff_a}{Proportional threshold for lowly expressed genes in cell
type A (range of [0-1]); for example, 0.2 means genes with some
expression in at least 20\% of cells are retained}

\item{cutoff_b}{Proportional expression threshold for cell type B (range of
[0-1])}

\item{pcg}{A character vector, contains the names of protein coding genes;
by default, uses the \code{pcg_human} data. This package also includes
\code{pcg_mouse}, but you can also use your own data}

\item{lrp}{A dataframe or matrix object with two columns, ligands names and
the names of their receptors; by default, uses the \code{lrp_human} data. This
package also includes \code{lrp_mouse}, but you can also use your own data}

\item{beta_max}{Upper limit of the test values of the PCSF objective
function parameter $I^2$, which is inversely proportional to the total
number of genes in a given cell-type pair; suggested to be 100 (default)
if the total number of genes in a given cell-type pair is above 10,000; if
the total number of genes is below 5,000, increase to 500}

\item{omega_min}{Start point of omega range; omega represents the edge cost
of the artificial network, but has been found to be less significant than
beta. Recommended minimum of \code{0.5}}

\item{omega_max}{End point of range between \code{omega_min} and \code{omega_max},
step size of \code{0.1}. Recommended maximum of \code{1.5}}

\item{depth}{Starting at each ligand-receptor pair in the resultant network,
how many steps out from that pair should be taken to generate each
neighborhood?}

\item{ntrial}{How many random network subsets shall be created to get an
empirical p-value for node prize and edge cost?}

\item{cores}{How many cores to use for parallel processing?}

\item{echo}{Should update messages be printed?}

\item{dir_out}{Folder used for output; if not specified, a "CytoTalk-output"
folder will be generated}
}
\value{
A list containing model parameters, prefential expression measure,
the integrated co-expression network, the results of the PCST, and resulting
pathways from the final extracted network
}
\description{
Main CytoTalk Pipeline
}
\examples{
{
cell_type_a <- "Macrophages"
cell_type_b <- "LuminalEpithelialCells"
cutoff_a <- 0.6
cutoff_b <- 0.6
# result <- CytoTalk::run_cytotalk(CytoTalk::scrna_cyto,
#                                  cell_type_a, cell_type_b,
#                                  cutoff_a, cutoff_b,
#                                  cores = 2)
}

}
