% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{subset_non_zero}
\alias{subset_non_zero}
\title{Subset Rows on Proportion Non-Zero}
\usage{
subset_non_zero(mat, cutoff)
}
\arguments{
\item{mat}{A numerical matrix}

\item{cutoff}{Threshold value in range [0, 1]; proportion of each row
required to be non-zero}
}
\value{
A matrix
}
\description{
Subset Rows on Proportion Non-Zero
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
cell_type_a <- "Macrophages"
cutoff_a <- 0.8
mat_a <- extract_group(cell_type_a, lst_scrna)
result <- subset_non_zero(mat_a, cutoff_a)
}

}
