% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{subset_rownames}
\alias{subset_rownames}
\title{Subset Rows on Rownames}
\usage{
subset_rownames(mat, labels)
}
\arguments{
\item{mat}{A numerical matrix}

\item{labels}{A subset of the matrix's rownames to filter by}
}
\value{
A matrix
}
\description{
Subset Rows on Rownames
}
\examples{
{
lst_scrna <- CytoTalk::scrna_cyto
pcg <- CytoTalk::pcg_human
cell_type_a <- "Macrophages"
mat_a <- extract_group(cell_type_a, lst_scrna)
result <- subset_rownames(mat_a, pcg)
}

}
